backend/api-gateway/requirements.txt
# API Gateway dependencies
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
httpx>=0.27.0
python-jose[cryptography]>=3.3.0
